Update sound config to globally disable/enable as per official
It's not technically implemented like official, but it's close enough based on our current implementation.

Officially (in 1.298) it's implemented so the setting applies on-demand, technically allowing for it to be changed live.
You can't actually change it live though, so effectively it doesn't really matter that we're just pre-emptively blocking it
from loading sounds.

Resolves #490