Update object NPC spawn & visibility behaviour (#590)
We intend to show objects on an opt-in basis, not hide them.
This is why there's a desync with the official client regarding things like the anvil.
The anvil isn't being sent in the NPC info packet, so it's not shown in the official client.
It's only shown here because the default visibility state is flipped.

The anvil isn't being sent to the client because it's excluded from spawning as an object NPC.
Supported types have been updated to allow it, so it's spawned and sent to the client now -- ensuring it's now displayed consistently in both the official and our client.

Additionally, officially we generate detailed meshes here, otherwise walking up to things like warp gates is a mess.

Refs #336