Server: Update CUser::CheckItemCount()
While it may appear a little wonky, this is official behaviour.
The main important change here is if the minimum required count is 0,
we typically consider it as successful (so long as it's a valid item).

This fixes the final issue preventing Abyss teleport gates from initially
triggering their menus.

Resolves #500