Strip N3Base wrapper for main solutions
The wrapper projects are needed for All.sln so they can be consistently built concurrently, but there's no real reason
to use the wrappers in Client.sln and Tools.sln.

Directly building the project means it can be loaded easily in the IDE (without having to change build configs) and
debugged.